server/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.1.13
idna==3.10
mysqlclient==2.2.6
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
